pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pratighat-ai"
version = "1.0.0"
description = "Indigenous AI-powered penetration testing and cyber defense platform"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Aman Rajak"}
]
keywords = ["security", "penetration-testing", "ai", "cybersecurity", "red-team"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "aiohttp>=3.9.0",
    "asyncio>=3.4.3",
    "langchain>=0.1.0",
    "langgraph>=0.0.20",
    "chromadb>=0.4.22",
    "ollama>=0.1.0",
    "rich>=13.7.0",
    "asciimatics>=1.15.0",
    "websockets>=12.0",
    "psutil>=5.9.0",
    "pyyaml>=6.0.1",
    "python-nmap>=0.7.1",
    "jinja2>=3.1.3",
    "markdown>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
pratighat = "pratighat.main:main"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
